Improved error handling for Bestiary
Errors from the database are now handled and will not result in a crash or infinite spinner. Error messages are also much more meaningful now. There is now an error CSS class for more visual indicators that something went wrong and error handling in the API route now has the same data format as a success for easier handling